refine oneflow compiler tests
